Add lookalike_urls/allowlist as a predefined entry
This CL adds the `lookalike_urls` relationship kind to the documentation. This is to allow site owners to list other sites they own so that browsers don't show security warnings because the URLs of the sites are similar looking.

For example, Chrome presently may show a "Did you mean example.com?" warning if the user visits "example1.com". If both sites are owned by the same owner, the owner might want to suppress the warning on example1.com.  Chrome plans to use Digital Asset Links to allow site owners to allowlist co-owned sites for lookalike warnings, and it will use the a new relationship type called `lookalike_urls/allowlist` for this purpose.